feat: component move MainPage component to its own file and clean up index.js
